spai_model/app/main.py
from fastapi import FastAPI, UploadFile, File, HTTPException
import torch
from torchvision import models, transforms
from torchvision.models import ResNet50_Weights
from PIL import Image
import io
import random

app = FastAPI()

# disease categories (ISIC 2019 classes)
categories = ["Melanoma", "Melanocytic Nevus", "Basal Cell Carcinoma", 
              "Actinic Keratosis", "Benign Keratosis", "Dermatofibroma", 
              "Squamous Cell Carcinoma", "Vascular Lesions"]

# typical probabilities for demo purposes
# (higher probability for benign conditions to avoid unnecessarily alarming users)
demo_probabilities = {
    "Melanocytic Nevus": 0.45,  # Common benign mole
    "Benign Keratosis": 0.25,   # Another common benign condition
    "Actinic Keratosis": 0.10,  # Pre-cancerous
    "Melanoma": 0.05,           # Dangerous skin cancer
    "Basal Cell Carcinoma": 0.05,
    "Dermatofibroma": 0.04,
    "Squamous Cell Carcinoma": 0.03,
    "Vascular Lesions": 0.03
}

# model loading for future use when you have trained weights
model = models.resnet50(weights=ResNet50_Weights.DEFAULT)
num_classes = len(categories)
model.fc = torch.nn.Linear(model.fc.in_features, num_classes)
model.eval()

# Image preprocessing - for validating images
transform = transforms.Compose([
    transforms.Resize((224, 224)),
    transforms.ToTensor(),
    transforms.Normalize([0.485, 0.456, 0.406], [0.229, 0.224, 0.225])
])

@app.post("/predict/")
async def predict(file: UploadFile = File(...)):
    try:
        # image to validate it's an actual image
        image_bytes = await file.read()
        image = Image.open(io.BytesIO(image_bytes)).convert("RGB")
        
        # validate image by attempting to transform it (ensures it's a valid image)
        _ = transform(image)
        
        # weighted random prediction instead of actual inference
        prediction = random.choices(
            population=list(demo_probabilities.keys()),
            weights=list(demo_probabilities.values()),
            k=1
        )[0]
        
        # confidence score and other predictions
        confidence = round(random.uniform(65, 95), 1)
        
        # list of other predictions with lower confidence
        all_predictions = []
        all_predictions.append({"class": prediction, "probability": confidence})
        
        # few other random predictions with lower confidence
        remaining = list(categories)
        if prediction in remaining:
            remaining.remove(prediction)
            
        for i in range(min(2, len(remaining))):
            other = random.choice(remaining)
            remaining.remove(other)
            other_confidence = round(random.uniform(5, 60), 1)
            all_predictions.append({"class": other, "probability": other_confidence})
        
        # Sort by probability
        all_predictions.sort(key=lambda x: x["probability"], reverse=True)
        
        return {
            "predicted_disease": prediction,
            "confidence": confidence,
            "all_predictions": all_predictions
        }
    
    except Exception as e:
        raise HTTPException(status_code=400, detail=f"Invalid image: {str(e)}")

@app.get("/")
async def root():
    return {"message": "Skin Disease Classification API"}